Updated BracketRandom
Simplified the math for finding values.

Corrected how the program checks to see if a value has already been used and stopped it from being run twice when unneeded using the selfExecuted variable.
Broke off the _alreadyUsedInBracket function for this purpose.

Added fill function to fill up a list with random, non-repeating values