BracketRandom Testing and Update
Created a test file for testing the speed and memory cost of BracketRandom. It turns out that BracketRandom is almost always the worst option and only ever can come in 2nd place sometimes under niche conditions. ChatGPT random is the best solution and brute force is second. I believe that the ChatGPT solution has worse memory efficiency, but I think my memory calculations are broken and I can't prove this besides writing the classes to files and seeing that BracketRandom is smaller. However, it's at the cost of a whole mega-ton of processing speed. Regardles, because I'm not going to take the time to implement the ChatGPT solution (and the fact that I dont think it would even hardly make a difference in time for what I'm doing) I will continue to use BracketRandom.

BracketRandom has been updated to include some support functions for the bedrock weather custom function.

Some errors in how bracket random works have been corrected.

BracketRandomAsList was created for testing but ended up being less efficient in every way compared to normal BracketRandom.