Path Formaliztion Update
Added the ability to formalize a path when getting it as a string rather than requiring the method to be called separately.

Formalizing a path will update the universal object, not just a temporary copy of the object, however, this is fine since there's no harm in formalizing a path.