BracketRandom Update
In [a recent commit](https://github.com/xLeviadeer/MCWiiU-Texture-Builder/commit/5c67cd64473a1befdcef29a126d2942d4e7edc0c) BracketRandom was tested and was found to be worse in basically every way compared to any other solution tested.

However, I was able to think of one redeeming quality, where the size of a written file as the data directly from BracketRandom will be smaller than other solutions. This could, really, just be converted to (the BracketRandom data structure) when writing and reading, but oh well. I'm not changing it now, however someone else can if they feel like it's a problem.

Because of this, I've natively added the ability to save and load BracketRandom.

In doing this I created the JsonWritable class which BracketRandom now extends. It's a framework of abstract and partial functions for writing a class as Json.